// libs/common/src/interfaces/category.ts
export interface RowCategoryCategoryId {
  category_id: number
}

export interface RowCategory {
  category_id: number
  category_name: string
  category_redirect_url: string
  category_domain_count: number
  category_redirect_type: string
}

export interface Category {
  categoryId: number
  categoryName: string
  categoryRedirectUrl: string
  categoryDmainCount: number
  categoryRedirectType: number
}